pyproject.toml
[project]
name = "ai-mental-therapist"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.136.0",
    "langchain>=0.3.27",
    "langchain-huggingface>=0.3.1",
    "langgraph>=0.6.8",
    "ollama>=0.6.1",
    "python-dotenv>=1.1.1",
    "pydantic>=2.13.3",
    "requests>=2.33.1",
    "twilio>=9.8.5",
    "uvicorn>=0.45.0",
    "langchain-openai>=1.1.16",
]
